refactor: update, insert 클래스 분리
