+++ README.md
@@ -3,7 +3,7 @@
 ## TODO
 - [x] property 파일로 db 연결 정보 관리
 - [x] update, insert 메서드 분리 
-- [ ] UpdateJdbcTemplate, InsertJdbcTemplate 클래스 분리
+- [x] UpdateJdbcTemplate, InsertJdbcTemplate 클래스 분리
 - [ ] UpdateJdbcTemplate, InsertJdbcTemplate 중복 추상클래스로 제거
 - [ ] SelectJdbcTemplate 클래스 분리
 - [ ] mapRow, setValues 인터페이스 분리
