+++ app/src/main/java/com/techcourse/dao/InsertJdbcTemplate.java
@@ -1,0 +1,42 @@
+package com.techcourse.dao;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+import javax.sql.DataSource;
+
+import com.techcourse.domain.User;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class InsertJdbcTemplate {
+
+    private static final Logger log = LoggerFactory.getLogger(InsertJdbcTemplate.class);
+
+    public void insert(User user, DataSource dataSource) {
+        final String sql = createQueryForInsert();
+
+        try (Connection conn = dataSource.getConnection();
+             PreparedStatement pstmt = conn.prepareStatement(sql)) {
+
+            log.debug("query : {}", sql);
+
+            setValuesForInsert(user, pstmt);
+            pstmt.executeUpdate();
+        } catch (SQLException e) {
+            log.error(e.getMessage(), e);
+            throw new RuntimeException(e);
+        }
+    }
+
+    private String createQueryForInsert() {
+        return "insert into users (account, password, email) values (?, ?, ?)";
+    }
+
+    private void setValuesForInsert(User user, PreparedStatement pstmt) throws SQLException {
+        pstmt.setString(1, user.getAccount());
+        pstmt.setString(2, user.getPassword());
+        pstmt.setString(3, user.getEmail());
+    }
+}
