+++ app/src/main/java/com/techcourse/dao/UpdateJdbcTemplate.java
@@ -1,0 +1,43 @@
+package com.techcourse.dao;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+import javax.sql.DataSource;
+
+import com.techcourse.domain.User;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class UpdateJdbcTemplate {
+
+    private static final Logger log = LoggerFactory.getLogger(UpdateJdbcTemplate.class);
+
+    public void update(User user, DataSource dataSource) {
+        final String sql = createQueryForUpdate();
+
+        try (Connection conn = dataSource.getConnection();
+             PreparedStatement pstmt = conn.prepareStatement(sql)) {
+
+            log.debug("query : {}", sql);
+
+            setValuesForUpdate(user, pstmt);
+            pstmt.executeUpdate();
+        } catch (SQLException e) {
+            log.error(e.getMessage(), e);
+            throw new RuntimeException(e);
+        }
+    }
+
+    private String createQueryForUpdate() {
+        return "update users set account = ?, password = ?, email = ? where id = ?";
+    }
+
+    private void setValuesForUpdate(User user, PreparedStatement pstmt) throws SQLException {
+        pstmt.setString(1, user.getAccount());
+        pstmt.setString(2, user.getPassword());
+        pstmt.setString(3, user.getEmail());
+        pstmt.setLong(4, user.getId());
+    }
+}
